New private module rand::seq::slice
